backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.30.0
bcrypt==4.1.2
certifi==2024.7.4
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.6.1
email_validator==2.1.2
exceptiongroup==1.2.2
fastapi==0.112.2
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.8
Jinja2==3.1.4
loguru==0.7.2
makefun==1.15.6
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pendulum==3.0.0
pip3-autoremove==1.2.2
pwdlib==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
redis==4.6.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
time-machine==2.15.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==13.0
